completed vid

+++ frontend/src/components/WorkoutDetails.js
@@ -1,17 +1,27 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext()
 
     const handleClick = async () => {
         // const response = await fetch('/api/workouts/' + workout._id, {
         //     method: "DELETE"
         // })
+
+        if (!user) {
+            return
+        }
+
         const response = await fetch(`${process.env.REACT_APP_API_URL}/api/workouts/${workout._id}`, {
-            method: "DELETE"
+            method: "DELETE",
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
